fix: resolve all 26 clippy warnings — 0 warnings now
Changes across 11 files:
- config.rs, download.rs, output.rs, pipeline.rs, main.rs: collapse nested if
- download.rs: map_or → is_some_and, sort_by → sort_by_key
- pipeline.rs: &PathBuf → &Path
- scene_detect.rs: refactor with_score to accept &ScoreDetails (10→6 params)
- frames/mod.rs: remove unnecessary cast and parens
- transcript.rs: while let → for loop
- error.rs: unwrap_or_else → unwrap_or
- setup.rs: manual split → split_once
- frames/timestamp.rs: map_or → is_some_and

cargo clippy: No issues found
cargo test: 263 passed

+++ src/frames/mod.rs
@@ -31,9 +31,7 @@ pub(crate) fn even_indices(count: usize, n: usize) -> Vec<usize> {
     if n <= 1 {
         return vec![0];
     }
-    (0..n)
-        .map(|i| (i * (count - 1) / (n - 1)) as usize)
-        .collect()
+    (0..n).map(|i| i * (count - 1) / (n - 1)).collect()
 }
 
 pub(crate) fn scale_filter(resolution: u32) -> String {
